[1274] Factor out test client builder
Remove unused method.
